refactor mapbox stuff into custom hook
